Altered Mission Statement
Made adjustments to direct people to relevant information for each
wording that may require more explanation. Also expanded shortened form
of words that may be open to ambiguity unnecessarily (spaces as opposed
to referring to Hackspace where it is the UK Hackspace Foundation of
Hackspace and not to be confused with Hackerspace or other definitions
of the word).  Also expanded wording to refer to ourselves as the UK
Hackspace Foundation in fulll as opposed to shortening to 'Foundation'
for clarity of which is being referred to. Adjusted the flow of the
paragraphs to intend to read as an introduction to the foundation and
terms, an explanation of action and an outro of future development,
wording was also adjusted to attempt to not purposefully exclude those
outside of the Hackspace definition and that this is open to alteration.

Also removed references to other files where those other files did not
elaborate on the word or wording linked from.